Cleanup and move stuff around a bit
