+++ Source/EventFlow.Tests/UnitTests/Specifications/TestSpecifications.cs
@@ -1,0 +1,60 @@
+﻿// The MIT License (MIT)
+//
+// Copyright (c) 2015 Rasmus Mikkelsen
+// https://github.com/rasmus/EventFlow
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy of
+// this software and associated documentation files (the "Software"), to deal in
+// the Software without restriction, including without limitation the rights to
+// use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of
+// the Software, and to permit persons to whom the Software is furnished to do so,
+// subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in all
+// copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS
+// FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
+// COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER
+// IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
+// CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+
+using System.Collections.Generic;
+using EventFlow.Specifications;
+
+namespace EventFlow.Tests.UnitTests.Specifications
+{
+    public static class TestSpecifications
+    {
+        public class IsAboveSpecification : Specification<int>
+        {
+            private readonly int _limit;
+
+            public IsAboveSpecification(
+                int limit)
+            {
+                _limit = limit;
+            }
+
+            protected override IEnumerable<string> IsNotStatisfiedBecause(int obj)
+            {
+                if (obj <= _limit)
+                {
+                    yield return $"{obj} is less or equal than {_limit}";
+                }
+            }
+        }
+
+        public class IsTrueSpecification : Specification<bool>
+        {
+            protected override IEnumerable<string> IsNotStatisfiedBecause(bool obj)
+            {
+                if (!obj)
+                {
+                    yield return "Its false!";
+                }
+            }
+        }
+    }
+}
