+++ Source/EventFlow.Tests/UnitTests/Specifications/TestSpecificationsTests.cs
@@ -1,0 +1,71 @@
+﻿// The MIT License (MIT)
+//
+// Copyright (c) 2015 Rasmus Mikkelsen
+// https://github.com/rasmus/EventFlow
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy of
+// this software and associated documentation files (the "Software"), to deal in
+// the Software without restriction, including without limitation the rights to
+// use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of
+// the Software, and to permit persons to whom the Software is furnished to do so,
+// subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in all
+// copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS
+// FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
+// COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER
+// IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
+// CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+
+using FluentAssertions;
+using NUnit.Framework;
+
+namespace EventFlow.Tests.UnitTests.Specifications
+{
+    public class TestSpecificationsTests
+    {
+        [Test]
+        public void IsTrueSpecification_ReturnsTrue_ForTrue()
+        {
+            // Arrange
+            var isTrue = new TestSpecifications.IsTrueSpecification();
+
+            // Act
+            var isSatisfiedBy = isTrue.IsSatisfiedBy(true);
+
+            // Act
+            isSatisfiedBy.Should().BeTrue();
+        }
+
+        [Test]
+        public void IsTrueSpecification_ReturnsFalse_ForFalse()
+        {
+            // Arrange
+            var isTrue = new TestSpecifications.IsTrueSpecification();
+
+            // Act
+            var isSatisfiedBy = isTrue.IsSatisfiedBy(false);
+
+            // Act
+            isSatisfiedBy.Should().BeFalse();
+        }
+
+        [TestCase(4, 3, false)]
+        [TestCase(4, 4, false)]
+        [TestCase(4, 5, true)]
+        public void IsAboveSpecification_Returns_Correct(int limit, int obj, bool expectedIsSatisfiedBy)
+        {
+            // Arrange
+            var isAbove = new TestSpecifications.IsAboveSpecification(limit);
+
+            // Act
+            var isSatisfiedBy = isAbove.IsSatisfiedBy(obj);
+
+            // Assert
+            isSatisfiedBy.Should().Be(expectedIsSatisfiedBy);
+        }
+    }
+}
